requirements.txt
# ============================================
# Antigravity - AI-Powered Health Assistant
# Requirements File
# ============================================

# Core Framework
streamlit>=1.30.0

# Deep Learning (Pneumonia Detection)
torch>=2.0.0
torchvision>=0.15.0
torchcam>=0.4.0

# Classical ML (Heart Disease)
scikit-learn>=1.3.0
joblib>=1.3.0

# Data & Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Image Processing
Pillow>=10.0.0

# PDF Report Generation
fpdf2>=2.7.0

# Explainability
shap>=0.42.0

# Utilities
requests>=2.31.0
